move coding history factory, etc to own namespace
