+++ Packager/Factories/CodingHistory/ICodingHistoryFactory.cs
@@ -1,0 +1,10 @@
+using Packager.Models.FileModels;
+using Packager.Models.PodMetadataModels;
+
+namespace Packager.Factories.CodingHistory
+{
+    public interface ICodingHistoryFactory
+    {
+        string Generate(AudioPodMetadata metadata, DigitalAudioFile provenance, AbstractFile model);
+    }
+}
