+++ Packager/Packager.csproj
@@ -88,14 +88,19 @@
     <Compile Include="Engine\EngineExitCodes.cs" />
     <Compile Include="Exceptions\UserCancelledException.cs" />
     <Compile Include="Extensions\XElmentExtensions.cs" />
+    <Compile Include="Factories\CodingHistory\AbstractCodingHistoryGenerator.cs" />
     <Compile Include="Factories\AbstractEmbeddedMetadataFactory.cs" />
     <Compile Include="Factories\AudioCarrierDataFactory.cs" />
-    <Compile Include="Factories\CodingHistoryFactory.cs" />
+    <Compile Include="Factories\CodingHistory\CodingHistoryFactory.cs" />
     <Compile Include="Factories\EmbeddedAudioMetadataFactory.cs" />
     <Compile Include="Factories\EmbeddedVideoMetadataFactory.cs" />
+    <Compile Include="Factories\CodingHistory\ICodingHistoryFactory.cs" />
+    <Compile Include="Factories\CodingHistory\ICodingHistoryGenerator.cs" />
     <Compile Include="Factories\IEmbeddedMetadataFactory.cs" />
     <Compile Include="Factories\ImportableFactory.cs" />
     <Compile Include="Factories\IPlaceHolderFactory.cs" />
+    <Compile Include="Factories\CodingHistory\LacquerOrCylinderCodingHistoryGenerator.cs" />
+    <Compile Include="Factories\CodingHistory\OpenReelCodingHistoryGenerator.cs" />
     <Compile Include="Factories\PlaceHolderFactory.cs" />
     <Compile Include="Factories\SettingsFactory.cs" />
     <Compile Include="Models\EmailMessageModels\SuccessEmailMessage.cs" />
